+++ crates/wdk-build/src/lib.rs
@@ -11,19 +11,19 @@
 //! models (WDM, KMDF, UMDF).
 
 #![cfg_attr(nightly_toolchain, feature(assert_matches))]
+use std::{fmt, str::FromStr};
 
 pub use bindgen::BuilderExt;
 use metadata::TryFromCargoMetadataError;
 
 pub mod cargo_make;
 pub mod metadata;
-/// Module for utility code related to the cargo-make experience for building
-/// drivers.
+
 pub mod utils;
 
 mod bindgen;
 
-use std::{env, path::PathBuf};
+use std::{env, path::PathBuf, sync::LazyLock};
 
 use cargo_metadata::MetadataCommand;
 use serde::{Deserialize, Serialize};
@@ -87,6 +87,28 @@ pub enum CpuArchitecture {
     Arm64,
 }
 
+impl FromStr for CpuArchitecture {
+    type Err = String;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s.to_lowercase().as_str() {
+            "amd64" => Ok(Self::Amd64),
+            "arm64" => Ok(Self::Arm64),
+            _ => Err(format!("'{s}' is not a valid target architecture")),
+        }
+    }
+}
+
+impl fmt::Display for CpuArchitecture {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let s = match self {
+            Self::Amd64 => "amd64",
+            Self::Arm64 => "arm64",
+        };
+        write!(f, "{s}")
+    }
+}
+
 /// The configuration parameters for KMDF drivers
 #[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq, Hash)]
 #[serde(
@@ -131,7 +153,35 @@ pub enum ConfigError {
         directory: String,
     },
 
-    /// Error returned when an
+    /// Error returned when a package is not found in Cargo metadata
+    #[error("cannot find wdk-build package in Cargo metadata")]
+    WdkBuildPackageNotFoundInCargoMetadata,
+
+    /// Error returned Cargo manifest contains an unsupported edition
+    #[error("Cargo manifest contains unsupported Rust edition: {edition}")]
+    UnsupportedRustEdition {
+        /// Edition of the Cargo manifest that was not supported
+        edition: String,
+    },
+
+    /// Error returned when `bindgen` does not support `rust-version` in Cargo
+    /// manifest
+    #[error("Rust version {msrv} not supported by Bindgen: {reason}")]
+    MsrvNotSupportedByBindgen {
+        /// MSRV that was not supported by Bindgen
+        msrv: String,
+        /// Reason why the MSRV was not supported
+        reason: String,
+    },
+
+    /// Error returned when `semver` parsing of the Rust version fails
+    #[error("failed to parse rust-version in manifest")]
+    RustVersionParseError {
+        /// [`semver::Error`] that caused parsing the Rust version to fail
+        #[source]
+        error_source: semver::Error,
+    },
+
     /// `utils::PathExt::strip_extended_length_path_prefix` operation fails
     #[error(transparent)]
     StripExtendedPathPrefixError(#[from] utils::StripExtendedPathPrefixError),
@@ -150,7 +200,7 @@ pub enum ConfigError {
 
     /// Error returned when the WDK version string does not match the expected
     /// format
-    #[error("The WDK version string provided ({version}) was not in a valid format.")]
+    #[error("the WDK version string provided ({version}) was not in a valid format")]
     WdkVersionStringFormatError {
         /// The incorrect WDK version string.
         version: String,
@@ -174,8 +224,14 @@ pub enum ConfigError {
     /// Error returned when the c runtime is not configured to be statically
     /// linked
     #[error(
-        "the C runtime is not properly configured to be statically linked. This is required for building \
-         WDK drivers. The recommended solution is to add the following snippiet to a `.config.toml` file: See https://doc.rust-lang.org/reference/linkage.html#static-and-dynamic-c-runtimes for more ways to enable static crt linkage."
+        "the C runtime is not properly configured to be statically linked. This is required for building WDK drivers. The recommended solution is to add the following snippet to a \
+        `.cargo/config.toml` file:
+[build]
+rustflags = [\"-C\", \"target-feature=+crt-static\"]
+
+\
+        See https://doc.rust-lang.org/reference/linkage.html#static-and-dynamic-c-runtimes for more ways \
+        to enable static crt linkage"
     )]
     StaticCrtNotEnabled,
 
@@ -185,8 +241,28 @@ pub enum ConfigError {
     SerdeError(#[from] metadata::Error),
 }
 
+/// Subset of APIs in the Windows Driver Kit
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub enum ApiSubset {
+    /// API subset typically required for all Windows drivers
+    Base,
+    /// API subset required for WDF (Windows Driver Framework) drivers: <https://learn.microsoft.com/en-us/windows-hardware/drivers/ddi/_wdf/>
+    Wdf,
+    /// API subset for GPIO (General Purpose Input/Output) drivers: <https://learn.microsoft.com/en-us/windows-hardware/drivers/ddi/_gpio/>
+    Gpio,
+    /// API subset for HID (Human Interface Device) drivers: <https://learn.microsoft.com/en-us/windows-hardware/drivers/ddi/_hid/>
+    Hid,
+    /// API subset for Parallel Ports drivers: <https://learn.microsoft.com/en-us/windows-hardware/drivers/ddi/_parports/>
+    ParallelPorts,
+    /// API subset for SPB (Serial Peripheral Bus) drivers: <https://learn.microsoft.com/en-us/windows-hardware/drivers/ddi/_spb/>
+    Spb,
+    /// API subset for Storage drivers: <https://learn.microsoft.com/en-us/windows-hardware/drivers/ddi/_storage/>
+    Storage,
+    /// API subset for USB (Universal Serial Bus) drivers: <https://learn.microsoft.com/en-us/windows-hardware/drivers/ddi/_usbref/>
+    Usb,
+}
+
 impl Default for Config {
-    #[must_use]
     fn default() -> Self {
         Self {
             wdk_content_root: utils::detect_wdk_content_root().expect(
@@ -248,7 +324,15 @@ impl Config {
         })
     }
 
-    fn emit_check_cfg_settings() {
+    /// Emit `cargo::rustc-check-cfg` directives corresponding to all the
+    /// possible `rustc-cfg` settings `wdk_build` could emit
+    ///
+    /// This is useful in situations where a library may not have a valid WDK
+    /// config available during build. This function is not needed if the build
+    /// was already configured via [`configure_wdk_binary_build`],
+    /// [`configure_wdk_library_build`], or
+    /// [`configure_wdk_library_build_and_then`]
+    pub fn emit_check_cfg_settings() {
         for (cfg_key, allowed_values) in EXPORTED_CFG_SETTINGS.iter() {
             let allowed_cfg_value_string =
                 allowed_values.iter().fold(String::new(), |mut acc, value| {
@@ -315,8 +399,7 @@ impl Config {
     ///
     /// This function will return an error if any of the required paths do not
     /// exist.
-    pub fn get_include_paths(&self) -> Result<Vec<PathBuf>, ConfigError> {
-        // FIXME: consider deprecating in favor of iter
+    pub fn include_paths(&self) -> Result<impl Iterator<Item = PathBuf>, ConfigError> {
         let mut include_paths = vec![];
 
         let include_directory = self.wdk_content_root.join("Include");
@@ -384,6 +467,21 @@ impl Config {
                         .canonicalize()?
                         .strip_extended_length_path_prefix()?,
                 );
+
+                // `ufxclient.h` relies on `ufxbase.h` being on the headers search path. The WDK
+                // normally does not automatically include this search path, but it is required
+                // here so that the headers can be processed successfully.
+                let ufx_include_path = km_or_um_include_path.join("ufx/1.1");
+                if !ufx_include_path.is_dir() {
+                    return Err(ConfigError::DirectoryNotFound {
+                        directory: ufx_include_path.to_string_lossy().into(),
+                    });
+                }
+                include_paths.push(
+                    ufx_include_path
+                        .canonicalize()?
+                        .strip_extended_length_path_prefix()?,
+                );
             }
             DriverConfig::Umdf(umdf_config) => {
                 let umdf_include_path = include_directory.join(format!(
@@ -403,7 +501,7 @@ impl Config {
             }
         }
 
-        Ok(include_paths)
+        Ok(include_paths.into_iter())
     }
 
     /// Return library include paths required to build and link based off of
@@ -415,7 +513,7 @@ impl Config {
     ///
     /// This function will return an error if any of the required paths do not
     /// exist.
-    pub fn get_library_paths(&self) -> Result<Vec<PathBuf>, ConfigError> {
+    pub fn library_paths(&self) -> Result<impl Iterator<Item = PathBuf>, ConfigError> {
         let mut library_paths = vec![];
 
         let library_directory = self.wdk_content_root.join("Lib");
@@ -490,14 +588,12 @@ impl Config {
         // Reverse order of library paths so that paths pushed later into the vec take
         // precedence
         library_paths.reverse();
-        Ok(library_paths)
+        Ok(library_paths.into_iter())
     }
 
     /// Return an iterator of strings that represent compiler definitions
     /// derived from the `Config`
-    pub fn get_preprocessor_definitions_iter(
-        &self,
-    ) -> impl Iterator<Item = (String, Option<String>)> {
+    pub fn preprocessor_definitions(&self) -> impl Iterator<Item = (String, Option<String>)> {
         // _WIN32_WINNT=$(WIN32_WINNT_VERSION);
         // WINVER=$(WINVER_VERSION);
         // WINNT=1;
@@ -536,10 +632,13 @@ impl Config {
         .chain(
             match self.driver_config {
                 DriverConfig::Wdm => {
-                    vec![]
+                    vec![
+                        ("_KERNEL_MODE", None), // Normally defined by msvc via /kernel flag
+                    ]
                 }
                 DriverConfig::Kmdf(kmdf_config) => {
                     let mut kmdf_definitions = vec![
+                        ("_KERNEL_MODE", None), // Normally defined by msvc via /kernel flag
                         ("KMDF_VERSION_MAJOR", Some(kmdf_config.kmdf_version_major)),
                         (
                             "KMDF_VERSION_MINOR",
@@ -603,8 +702,8 @@ impl Config {
             "--warn-=no-ignored-attributes",
             "--warn-=no-ignored-pragma-intrinsic",
             "--warn-=no-visibility",
-            "--warn-=no-microsoft-anon-tag",
-            "--warn-=no-microsoft-enum-forward-reference",
+            "--warn-=no-switch",
+            "--warn-=no-comment",
             // Don't warn for deprecated declarations. Deprecated items should be explicitly
             // blocklisted (i.e. by the bindgen invocation). Any non-blocklisted function
             // definitions will trigger a -WDeprecated warning
@@ -613,9 +712,233 @@ impl Config {
             // `_variable` are separate tokens already, and don't need `##` to concatenate
             // them)
             "--warn-=no-invalid-token-paste",
+            // Windows SDK & DDK headers rely on Microsoft extensions to C/C++
+            "--warn-=no-microsoft",
         ]
         .into_iter()
         .map(std::string::ToString::to_string)
+    }
+
+    /// Returns a [`String`] iterator over all the headers for a given
+    /// [`ApiSubset`]
+    ///
+    /// The iterator considers both the [`ApiSubset`] and the [`Config`] to
+    /// determine which headers to yield
+    pub fn headers(&self, api_subset: ApiSubset) -> impl Iterator<Item = String> {
+        match api_subset {
+            ApiSubset::Base => self.base_headers(),
+            ApiSubset::Wdf => self.wdf_headers(),
+            ApiSubset::Gpio => self.gpio_headers(),
+            ApiSubset::Hid => self.hid_headers(),
+            ApiSubset::ParallelPorts => self.parallel_ports_headers(),
+            ApiSubset::Spb => self.spb_headers(),
+            ApiSubset::Storage => self.storage_headers(),
+            ApiSubset::Usb => self.usb_headers(),
+        }
+        .into_iter()
+        .map(str::to_string)
+    }
+
+    fn base_headers(&self) -> Vec<&'static str> {
+        match &self.driver_config {
+            DriverConfig::Wdm | DriverConfig::Kmdf(_) => {
+                vec!["ntifs.h", "ntddk.h", "ntstrsafe.h"]
+            }
+            DriverConfig::Umdf(_) => {
+                vec!["windows.h"]
+            }
+        }
+    }
+
+    fn wdf_headers(&self) -> Vec<&'static str> {
+        if matches!(
+            self.driver_config,
+            DriverConfig::Kmdf(_) | DriverConfig::Umdf(_)
+        ) {
+            vec!["wdf.h"]
+        } else {
+            vec![]
+        }
+    }
+
+    fn gpio_headers(&self) -> Vec<&'static str> {
+        let mut headers = vec!["gpio.h"];
+        if matches!(self.driver_config, DriverConfig::Kmdf(_)) {
+            headers.extend(["gpioclx.h"]);
+        }
+        headers
+    }
+
+    fn hid_headers(&self) -> Vec<&'static str> {
+        let mut headers = vec!["hidclass.h", "hidsdi.h", "hidpi.h", "vhf.h"];
+        if matches!(
+            self.driver_config,
+            DriverConfig::Wdm | DriverConfig::Kmdf(_)
+        ) {
+            headers.extend(["hidpddi.h", "hidport.h", "kbdmou.h", "ntdd8042.h"]);
+        }
+
+        if matches!(self.driver_config, DriverConfig::Kmdf(_)) {
+            headers.extend(["HidSpiCx/1.0/hidspicx.h"]);
+        }
+        headers
+    }
+
+    fn parallel_ports_headers(&self) -> Vec<&'static str> {
+        let mut headers = vec!["ntddpar.h", "ntddser.h"];
+        if matches!(
+            self.driver_config,
+            DriverConfig::Wdm | DriverConfig::Kmdf(_)
+        ) {
+            headers.extend(["parallel.h"]);
+        }
+        headers
+    }
+
+    fn spb_headers(&self) -> Vec<&'static str> {
+        let mut headers = vec!["spb.h", "reshub.h"];
+        if matches!(
+            self.driver_config,
+            DriverConfig::Wdm | DriverConfig::Kmdf(_)
+        ) {
+            headers.extend(["pwmutil.h"]);
+        }
+        if matches!(self.driver_config, DriverConfig::Kmdf(_)) {
+            headers.extend(["spb/1.1/spbcx.h"]);
+        }
+        headers
+    }
+
+    fn storage_headers(&self) -> Vec<&'static str> {
+        let mut headers = vec![
+            "ehstorioctl.h",
+            "ntddcdrm.h",
+            "ntddcdvd.h",
+            "ntdddisk.h",
+            "ntddmmc.h",
+            "ntddscsi.h",
+            "ntddstor.h",
+            "ntddtape.h",
+            "ntddvol.h",
+            "ufs.h",
+        ];
+        if matches!(
+            self.driver_config,
+            DriverConfig::Wdm | DriverConfig::Kmdf(_)
+        ) {
+            headers.extend([
+                "mountdev.h",
+                "mountmgr.h",
+                "ntddchgr.h",
+                "ntdddump.h",
+                "storduid.h",
+                "storport.h",
+            ]);
+        }
+        if matches!(self.driver_config, DriverConfig::Kmdf(_)) {
+            headers.extend(["ehstorbandmgmt.h"]);
+        }
+        headers
+    }
+
+    fn usb_headers(&self) -> Vec<&'static str> {
+        let mut headers = vec![
+            "usb.h",
+            "usbfnbase.h",
+            "usbioctl.h",
+            "usbspec.h",
+            "Usbpmapi.h",
+        ];
+
+        if matches!(
+            self.driver_config,
+            DriverConfig::Wdm | DriverConfig::Kmdf(_)
+        ) {
+            headers.extend(["usbbusif.h", "usbdlib.h", "usbfnattach.h", "usbfnioctl.h"]);
+        }
+
+        if matches!(
+            self.driver_config,
+            DriverConfig::Kmdf(_) | DriverConfig::Umdf(_)
+        ) {
+            headers.extend(["wdfusb.h"]);
+        }
+
+        if matches!(self.driver_config, DriverConfig::Kmdf(_)) {
+            headers.extend([
+                "ucm/1.0/UcmCx.h",
+                "UcmTcpci/1.0/UcmTcpciCx.h",
+                "UcmUcsi/1.0/UcmucsiCx.h",
+                "ucx/1.6/ucxclass.h",
+                "ude/1.1/UdeCx.h",
+                "ufx/1.1/ufxbase.h",
+                "ufxproprietarycharger.h",
+                "urs/1.0/UrsCx.h",
+            ]);
+
+            if Self::should_include_ufxclient() {
+                headers.extend(["ufx/1.1/ufxclient.h"]);
+            }
+        }
+        headers
+    }
+
+    /// Determines whether to include the ufxclient.h header based on the Clang
+    /// version used by bindgen.
+    ///
+    /// The ufxclient.h header contains FORCEINLINE annotations that are invalid
+    /// according to the C standard. While MSVC silently ignores these in C
+    /// mode, older versions of Clang (pre-20.0) will error, even with MSVC
+    /// compatibility enabled.
+    ///
+    /// This function checks if the current Clang version is 20.0 or newer,
+    /// where the issue was fixed. See
+    /// <https://github.com/llvm/llvm-project/issues/124869> for details.
+    fn should_include_ufxclient() -> bool {
+        const MINIMUM_CLANG_MAJOR_VERISON_WITH_INVALID_INLINE_FIX: u32 = 20;
+
+        let clang_version = ::bindgen::clang_version();
+        match clang_version.parsed {
+            Some((major, _minor))
+                if major >= MINIMUM_CLANG_MAJOR_VERISON_WITH_INVALID_INLINE_FIX =>
+            {
+                true
+            }
+            Some(_) => {
+                tracing::info!(
+                    "Skipping ufxclient.h due to FORCEINLINE bug in {}",
+                    clang_version.full
+                );
+                false
+            }
+            None => {
+                tracing::warn!(
+                    "Failed to parse semver Major and Minor components from full Clang version \
+                     string: {}",
+                    clang_version.full
+                );
+                false
+            }
+        }
+    }
+
+    /// Returns a [`String`] containing the contents of a header file designed
+    /// for [`bindgen`](https://docs.rs/bindgen) to process
+    ///
+    /// The contents contain `#include`'ed headers based off the [`ApiSubset`]
+    /// and [`Config`], as well as any additional definitions required for the
+    /// headers to be processed successfully
+    pub fn bindgen_header_contents(
+        &self,
+        api_subsets: impl IntoIterator<Item = ApiSubset>,
+    ) -> String {
+        api_subsets
+            .into_iter()
+            .flat_map(|api_subset| {
+                self.headers(api_subset)
+                    .map(|header| format!("#include \"{header}\"\n"))
+            })
+            .collect::<String>()
     }
 
     /// Configure a Cargo build of a library that depends on the WDK. This
@@ -659,20 +982,27 @@ impl Config {
     ///
     /// This function will return an error if:
     /// * any of the required WDK paths do not exist
-    /// * the C runtime is not configured to be statically linked
+    /// * the C runtime is not configured to be statically linked for a
+    ///   kernel-mode driver
     ///
     /// # Panics
     ///
     /// Panics if the invoked from outside a Cargo build environment
     pub fn configure_binary_build(&self) -> Result<(), ConfigError> {
         if !Self::is_crt_static_linked() {
-            return Err(ConfigError::StaticCrtNotEnabled);
+            cfg_if::cfg_if! {
+                if #[cfg(all(wdk_build_unstable, skip_umdf_static_crt_check))] {
+                    if !matches!(self.driver_config, DriverConfig::Umdf(_)) {
+                        return Err(ConfigError::StaticCrtNotEnabled);
+                    }
+                } else {
+                    return Err(ConfigError::StaticCrtNotEnabled);
+                }
+            };
         }
 
-        let library_paths: Vec<PathBuf> = self.get_library_paths()?;
-
         // Emit linker search paths
-        for path in library_paths {
+        for path in self.library_paths()? {
             println!("cargo::rustc-link-search={}", path.display());
         }
 
@@ -684,6 +1014,12 @@ impl Config {
                 println!("cargo::rustc-link-lib=static=hal");
                 println!("cargo::rustc-link-lib=static=wmilib");
 
+                // Emit ARM64-specific libraries to link to derived from
+                // WindowsDriver.arm64.props
+                if self.cpu_architecture == CpuArchitecture::Arm64 {
+                    println!("cargo::rustc-link-lib=static=arm64rt");
+                }
+
                 // Linker arguments derived from WindowsDriver.KernelMode.props in Ni(22H2) WDK
                 println!("cargo::rustc-cdylib-link-arg=/DRIVER");
                 println!("cargo::rustc-cdylib-link-arg=/NODEFAULTLIB");
@@ -693,6 +1029,15 @@ impl Config {
                 // Linker arguments derived from WindowsDriver.KernelMode.WDM.props in Ni(22H2)
                 // WDK
                 println!("cargo::rustc-cdylib-link-arg=/ENTRY:DriverEntry");
+
+                // Ignore `LNK4257: object file was not compiled for kernel mode; the image
+                // might not run` since `rustc` has no support for `/KERNEL`
+                println!("cargo::rustc-cdylib-link-arg=/IGNORE:4257");
+
+                // Ignore `LNK4216: Exported entry point DriverEntry` since Rust currently
+                // provides no way to set a symbol's name without also exporting the symbol:
+                // https://github.com/rust-lang/rust/issues/67399
+                println!("cargo::rustc-cdylib-link-arg=/IGNORE:4216");
             }
             DriverConfig::Kmdf(_) => {
                 // Emit KMDF-specific libraries to link to
@@ -703,6 +1048,12 @@ impl Config {
                 println!("cargo::rustc-link-lib=static=WdfLdr");
                 println!("cargo::rustc-link-lib=static=WdfDriverEntry");
 
+                // Emit ARM64-specific libraries to link to derived from
+                // WindowsDriver.arm64.props
+                if self.cpu_architecture == CpuArchitecture::Arm64 {
+                    println!("cargo::rustc-link-lib=static=arm64rt");
+                }
+
                 // Linker arguments derived from WindowsDriver.KernelMode.props in Ni(22H2) WDK
                 println!("cargo::rustc-cdylib-link-arg=/DRIVER");
                 println!("cargo::rustc-cdylib-link-arg=/NODEFAULTLIB");
@@ -712,6 +1063,10 @@ impl Config {
                 // Linker arguments derived from WindowsDriver.KernelMode.KMDF.props in
                 // Ni(22H2) WDK
                 println!("cargo::rustc-cdylib-link-arg=/ENTRY:FxDriverEntry");
+
+                // Ignore `LNK4257: object file was not compiled for kernel mode; the image
+                // might not run` since `rustc` has no support for `/KERNEL`
+                println!("cargo::rustc-cdylib-link-arg=/IGNORE:4257");
             }
             DriverConfig::Umdf(umdf_config) => {
                 // Emit UMDF-specific libraries to link to
@@ -774,7 +1129,6 @@ impl From<DeserializableDriverConfig> for DriverConfig {
 }
 
 impl Default for KmdfConfig {
-    #[must_use]
     fn default() -> Self {
         // FIXME: determine default values from TargetVersion and _NT_TARGET_VERSION
         Self {
@@ -794,7 +1148,6 @@ impl KmdfConfig {
 }
 
 impl Default for UmdfConfig {
-    #[must_use]
     fn default() -> Self {
         // FIXME: determine default values from TargetVersion and _NT_TARGET_VERSION
         Self {
@@ -825,8 +1178,6 @@ impl CpuArchitecture {
     }
 
     /// Converts from a cargo-provided [`std::str`] to a [`CpuArchitecture`].
-    ///
-    /// #
     #[must_use]
     pub fn try_from_cargo_str<S: AsRef<str>>(cargo_str: S) -> Option<Self> {
         // Specifically not using the [`std::convert::TryFrom`] trait to be more
@@ -878,7 +1229,7 @@ pub fn find_top_level_cargo_manifest() -> PathBuf {
 ///
 /// Cargo build graphs that have no valid WDK configurations will emit a
 /// warning, but will still return [`Ok`]. This allows libraries
-/// designed for multiple configurations to sucessfully compile when built in
+/// designed for multiple configurations to successfully compile when built in
 /// isolation.
 ///
 /// # Errors
@@ -917,7 +1268,7 @@ pub fn configure_wdk_library_build() -> Result<(), ConfigError> {
 ///
 /// Cargo build graphs that have no valid WDK configurations will emit a
 /// warning, but will still return [`Ok`]. This allows libraries
-/// designed for multiple configurations to sucessfully compile when built in
+/// designed for multiple configurations to successfully compile when built in
 /// isolation.
 ///
 /// # Errors
@@ -967,13 +1318,53 @@ pub fn configure_wdk_binary_build() -> Result<(), ConfigError> {
     Config::from_env_auto()?.configure_binary_build()
 }
 
-// This currently only exports the driver type, but may export more metadata in
-// the future. `EXPORTED_CFG_SETTINGS` is a mapping of cfg key to allowed cfg
-// values
-lazy_static::lazy_static! {
-    // FIXME: replace lazy_static with std::Lazy once available: https://github.com/rust-lang/rust/issues/109736
-    static ref EXPORTED_CFG_SETTINGS: Vec<(&'static str, Vec<&'static str>)> =
-        vec![("DRIVER_MODEL-DRIVER_TYPE", vec!["WDM", "KMDF", "UMDF"])];
+/// This currently only exports the driver type, but may export more metadata in
+/// the future. `EXPORTED_CFG_SETTINGS` is a mapping of cfg key to allowed cfg
+/// values
+static EXPORTED_CFG_SETTINGS: LazyLock<Vec<(&'static str, Vec<&'static str>)>> =
+    LazyLock::new(|| vec![("DRIVER_MODEL-DRIVER_TYPE", vec!["WDM", "KMDF", "UMDF"])]);
+
+/// Detects the WDK build number.
+///
+/// Detects the Windows Driver Kit (WDK) build number by locating
+/// the WDK content root, retrieving the latest Windows SDK version, validating
+/// the version format, and extracting the build number.
+///
+/// # Returns
+///
+/// Returns a `Result<u32, ConfigError>`, which contains the WDK
+/// build number on success or a `ConfigError` on failure.
+///
+/// # Errors
+///
+/// Returns an error if:
+/// * The WDK content root cannot be detected.
+/// * The latest Windows SDK version cannot be retrieved.
+/// * The WDK version string format is invalid.
+/// * The WDK version number cannot be parsed.
+///
+/// # Panics
+///
+/// Panics if the WDK version number cannot be extracted from
+/// the version string.
+pub fn detect_wdk_build_number() -> Result<u32, ConfigError> {
+    let wdk_content_root =
+        utils::detect_wdk_content_root().ok_or(ConfigError::WdkContentRootDetectionError)?;
+    let detected_sdk_version =
+        utils::get_latest_windows_sdk_version(&wdk_content_root.join("Lib"))?;
+
+    if !utils::validate_wdk_version_format(&detected_sdk_version) {
+        return Err(ConfigError::WdkVersionStringFormatError {
+            version: detected_sdk_version,
+        });
+    }
+
+    let wdk_build_number =
+        str::parse::<u32>(&utils::get_wdk_version_number(&detected_sdk_version)?).unwrap_or_else(
+            |_| panic!("Couldn't parse WDK version number! Version number: {detected_sdk_version}"),
+        );
+
+    Ok(wdk_build_number)
 }
 
 #[cfg(test)]
@@ -1156,6 +1547,66 @@ mod tests {
         assert_eq!(CpuArchitecture::try_from_cargo_str("arm"), None);
     }
 
+    mod bindgen_header_contents {
+        use super::*;
+        use crate::{KmdfConfig, UmdfConfig};
+
+        #[test]
+        fn wdm() {
+            let config = with_env(&[("CARGO_CFG_TARGET_ARCH", "x86_64")], || Config {
+                driver_config: DriverConfig::Wdm,
+                ..Default::default()
+            });
+
+            assert_eq!(
+                config.bindgen_header_contents([ApiSubset::Base]),
+                r#"#include "ntifs.h"
+#include "ntddk.h"
+#include "ntstrsafe.h"
+"#,
+            );
+        }
+
+        #[test]
+        fn kmdf() {
+            let config = with_env(&[("CARGO_CFG_TARGET_ARCH", "x86_64")], || Config {
+                driver_config: DriverConfig::Kmdf(KmdfConfig {
+                    kmdf_version_major: 1,
+                    target_kmdf_version_minor: 33,
+                    minimum_kmdf_version_minor: None,
+                }),
+                ..Default::default()
+            });
+
+            assert_eq!(
+                config.bindgen_header_contents([ApiSubset::Base, ApiSubset::Wdf]),
+                r#"#include "ntifs.h"
+#include "ntddk.h"
+#include "ntstrsafe.h"
+#include "wdf.h"
+"#,
+            );
+        }
+
+        #[test]
+        fn umdf() {
+            let config = with_env(&[("CARGO_CFG_TARGET_ARCH", "aarch64")], || Config {
+                driver_config: DriverConfig::Umdf(UmdfConfig {
+                    umdf_version_major: 2,
+                    target_umdf_version_minor: 15,
+                    minimum_umdf_version_minor: None,
+                }),
+                ..Default::default()
+            });
+
+            assert_eq!(
+                config.bindgen_header_contents([ApiSubset::Base, ApiSubset::Wdf]),
+                r#"#include "windows.h"
+#include "wdf.h"
+"#,
+            );
+        }
+    }
     mod compute_wdffunctions_symbol_name {
         use super::*;
         use crate::{KmdfConfig, UmdfConfig};